refactor : change RacingGame's package
